requirements.txt
discord.py==2.6.4
numpy==2.4.2
transformers==5.2.0
